Refactor details list to list description
Since there has already been implemented a details list we need to
refactor the new one.
