+++ base.scss
@@ -44,7 +44,7 @@
 @import "./src/stories/Library/Modals/modal-pause/modal-pause";
 
 @import "./src/stories/Blocks/material-manifestation-item/material-manifestation-item";
-@import "./src/stories/Library/details-list/details-list";
+@import "./src/stories/Library/Lists/list-description/list-description";
 
 // Blocks
 @import "./src/stories/Blocks/footer/footer";
