+++ src/stories/Library/Lists/list-description/list-description.scss
@@ -1,4 +1,4 @@
-dl.details-list {
+dl.list-description {
   @extend %text-small-caption;
 
   color: $c-text-secondary-gray;
